# apiDoc.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: myDeben
  description: "API Documentation"
  contact:
    email: "youssuf.radi@deemalab.com"
host: "localhost:3000"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /user/profile:
    get:
      tags:
      - "User"
      summary: "Returns all user details and orders"
      description: "View Full details:
                    - Email
                    - Name
                    - Checked In store
                    - Orders Total Amount
                    - Order History"
      operationId: "userProfile"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Full Profile"
          schema:
            $ref: "#/definitions/Profile"
        400:
          description: "Wrong User token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /user/discover:
    get:
      tags:
      - "User"
      summary: "Returns all gems provided by the store"
      description: "User can see the main services or recommendations by hotel admin"
      operationId: "userDiscover"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "All gems"
          schema:
            type: "object"
            properties:
              gems:
                type: "array"
                items:
                  $ref: "#/definitions/Gem"
        400:
          description: "Wrong User token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /user/order:
    get:
      tags:
      - "User"
      summary: "Returns all orders made by the user"
      description: "User coulde see history of my orders and track current status "
      operationId: "userOrder"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "All orders"
          schema:
            type: "object"
            properties:
              oredrs:
                type: "array"
                items:
                  $ref: "#/definitions/Order"
        400:
          description: "Wrong User token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
    post:
      tags:
      - "User"
      summary: "User Making Orders"
      description: "User can pick a certain category then menu item and order"
      operationId: "userAddOrder"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Takes order of a user to order"
        required: true
        schema:
          type: "object"
          required:
          - "store_id"
          - "order"
          properties:
            store_id:
              type: "integer"
            order:
              type: "array"
              items:
                type: "object"
                properties:
                  quantity:
                    type: "integer"
                  product_id:
                    type: "integer"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Ordered Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong User token Used or store doesnot exist"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /user/checkin:
    post:
      tags:
      - "User"
      summary: "User Check in to store using QR"
      description: "Travellers can scan QR Code once he arrives at hotel so that 
                he can checkin the hotel services on Deben
                Output from the QR code is sent to the server for checkin"
      operationId: "checkInQR"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Checks user into a certain store with id and name"
        required: true
        schema:
          type: "object"
          required:
          - "store_name"
          - "store_id"
          properties:
            store_name:
              type: "string"
            store_id:
              type: "integer"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Checked In Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong User token Used or Already checked_in"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Store doesnot exist"
  /user/checkin/{token}:
    get:
      tags:
      - "User"
      summary: "User Check in to store using code"
      description: "Travellers can enter code given by hotel manager once he arrives at hotel so that 
                he can checkin the hotel services on Deben"
      operationId: "checkInCode"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "token"
        description: "Token given by hotel admin for user to checkin"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Checked In Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong User token Used or Already checked_in"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Token doesnot exist"
  /user/category:
    get:
      tags:
      - "User"
      summary: "Retrives all store categories for current checked in store"
      description: "Shows all available categories for the store user is currently checked in with all details"
      operationId: "userShowCategories"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Array of categories"
          schema:
            type: "object"
            properties:
              categories:
                type: "array"
                items:
                  $ref: "#/definitions/Category"
        400:
          description: "Wrong User token Used / User didnot previously check in"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /user/product:
    get:
      tags:
      - "User"
      summary: "Retrives all store category products for current checked in store"
      description: "Shows all available category products for the store user is currently checked in with all details"
      operationId: "userShowProducts"
      produces:
      - "application/json"
      parameters:
      - name: "categoryId"
        in: "query"
        description: "Category Id to retreive products inside"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [user]
      responses:
        200:
          description: "Array of Products"
          schema:
            type: "object"
            properties:
              products:
                type: "array"
                items:
                  $ref: "#/definitions/Product"
        400:
          description: "Wrong User token Used / User didnot previously check in"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Category Id not found"
...
definitions:
  Success:
    type: "object"
    properties:
      detail:
        type: "string" 
      success:
        type: "boolean"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      email:
        type: "string" 
        format: "email"
      name:
        type: "string"
      checkin_store_name:
        type: "string" 
      checkin_store_id:
        type: "integer"
  Order:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
      price:
        type: "integer"
      total_price:
        type: "integer"
      quantity:
        type: "integer"
      store_id:
        type: "integer"
      user_id:
        type: "integer"
      checked_out:
        type: "boolean"
      served:
        type: "boolean"
  Profile:
    type: "object"
    properties:
      user:
        $ref: "#/definitions/User"
      total:
        type: "integer"
      orders:
        type: "array"
        items:
          $ref: "#/definitions/Order"
  Gem:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      price:
        type: "integer"
      gem:
        type: "boolean"
      picture:
        type: "string"
      store_id:
        type: "integer"
      description:
        type: "string"
      category_id:
        type: "integer"
      quantity:
        type: "integer"
      user_id:
        type: "integer"
      checked_out:
        type: "boolean"
  Category:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
  Product:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
      price:
        type: "integer"
      gem:
        type: "boolean"
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    in: header
    name: "x-access-token"